requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.0
python-jose[cryptography]>=3.3.0
certifi>=2024.2.2
pyyaml>=6.0.1
python-multipart>=0.0.9
openai>=1.30.0
python-dotenv>=1.0.0